Allow the option to use custom challengers
When an authentication request is received, pick out the
challenger plugin from the path and use that in the
authenticate method instead of the hard-coded defaults.

This commit does not alter the behaviour of the command
auth on the command line.  Some thoughts on how this
should be altered need to happen...